Fixes broken step-up talent.
The were 2 issues to the step-up talent, first one was that it added 2x config value, and second one was that on world switch it kept increasing the step-up value by config value.

It is fixed by removing unnecessary sum in the StepTalent#onAdded method and by changing player.stepHeight to 1.0F (default value) in StepTalent#onEntityCreated method.

Fixes of #255