Fix a bug that allowed creating portals in any dimension.
By the code it is assumed that portals will be created only in overworld, however, there were no additional checks on that, and players were able to create portals in any dimension, even in Vault Dimension.

This is a major bug, as players could always create exit portal at any position in the vault.

Fixes #273 and #281.